requirements.txt
Django==4.2.*
graphene-django==3.2.0
django-filter==23.5
typing-extensions==4.9.0
django-settings-holder==0.1.2
loguru==0.7.*
pgvector==0.2.*
psycopg2-binary==2.9.*
django-taggit==4.0.*
django-model-utils==4.2.*

# Testing
pytest==8.0.1
coverage==7.4.2
pytest-django==4.8.0
pre-commit==3.6.2
tox==4.13.0
tox-gh-actions==3.2.0
faker==23.2.1
sqlparse==0.4.4
django-graphiql-debug-toolbar==0.2.0
ipython
